+++ src/skins-qt/widget.h
@@ -38,8 +38,8 @@ protected:
     void add_input (int width, int height, bool track_motion, bool drawable);
     void add_drawable (int width, int height);
 
-    void set_scale (int scale) { m_scale = scale; }
-    void resize (int w, int h) { QWidget::resize (w * m_scale, h * m_scale); }
+    void set_scale (double scale) { m_scale = scale; }
+    void resize (int w, int h) { QWidget::resize ((int) (w * m_scale), (int) (h * m_scale)); }
 
 #ifdef Q_OS_MAC
     /* repaint() causes graphical glitches on OS X
@@ -77,7 +77,7 @@ private:
         { event->setAccepted (close ()); }
 
     bool m_drawable = false;
-    int m_scale = 1;
+    double m_scale = 1.0;
 };
 
 #endif // SKINS_WIDGET_H
